per Matt move ReadOnlyGoal to its own comp
